+++ packages/backend/src/server/api/endpoints/i/update.ts
@@ -186,6 +186,7 @@ export const paramDef = {
 		isCat: { type: 'boolean' },
 		injectFeaturedNote: { type: 'boolean' },
 		receiveAnnouncementEmail: { type: 'boolean' },
+		receiveEmojiRequestResultEmail: { type: 'boolean' },
 		alwaysMarkNsfw: { type: 'boolean' },
 		autoSensitive: { type: 'boolean' },
 		followingVisibility: { type: 'string', enum: ['public', 'followers', 'private'] },
@@ -360,6 +361,7 @@ export default class extends Endpoint<typeof meta, typeof paramDef> { // eslint-
 			if (typeof ps.isCat === 'boolean') updates.isCat = ps.isCat;
 			if (typeof ps.injectFeaturedNote === 'boolean') profileUpdates.injectFeaturedNote = ps.injectFeaturedNote;
 			if (typeof ps.receiveAnnouncementEmail === 'boolean') profileUpdates.receiveAnnouncementEmail = ps.receiveAnnouncementEmail;
+			if (typeof ps.receiveEmojiRequestResultEmail === 'boolean') profileUpdates.receiveEmojiRequestResultEmail = ps.receiveEmojiRequestResultEmail;
 			if (typeof ps.alwaysMarkNsfw === 'boolean') {
 				policies ??= await this.roleService.getUserPolicies(user.id);
 				if (policies.alwaysMarkNsfw) throw new ApiError(meta.errors.restrictedByRole);